requirements.txt
Django~=4.1.3
djangorestframework~=3.14.0
django-debug-toolbar~=3.7.0